+++ rapier2d/src/lib.rs
@@ -3,7 +3,6 @@
 //! In addition to the instruction in bevy-tnua's documentation:
 //!
 //! * Add [`TnuaRapier2dPlugin`] to the Bevy app.
-//! * Add [`TnuaRapier2dIOBundle`] to each character entity controlled by Tnua.
 //! * Optionally: Add [`TnuaRapier2dSensorShape`] to either entity of the character controller by
 //!   Tnua or to the to the sensor entities. If exists, the shape on the sensor entity overrides
 //!   the one on the character entity for that specific sensor.
@@ -17,17 +16,16 @@ use bevy_rapier2d::rapier;
 use bevy_rapier2d::rapier::prelude::InteractionGroups;
 use bevy_rapier2d::{parry, prelude::*};
 
-use bevy_tnua_physics_integration_layer::data_for_backends::TnuaGhostSensor;
+use bevy_tnua_physics_integration_layer::TnuaPipelineSystems;
+use bevy_tnua_physics_integration_layer::TnuaSystems;
 use bevy_tnua_physics_integration_layer::data_for_backends::TnuaGravity;
 use bevy_tnua_physics_integration_layer::data_for_backends::TnuaToggle;
 use bevy_tnua_physics_integration_layer::data_for_backends::{TnuaGhostPlatform, TnuaNotPlatform};
+use bevy_tnua_physics_integration_layer::data_for_backends::{TnuaGhostSensor, TnuaSensorOf};
 use bevy_tnua_physics_integration_layer::data_for_backends::{
     TnuaMotor, TnuaProximitySensor, TnuaProximitySensorOutput, TnuaRigidBodyTracker,
 };
 use bevy_tnua_physics_integration_layer::obstacle_radar::TnuaObstacleRadar;
-use bevy_tnua_physics_integration_layer::subservient_sensors::TnuaSubservientSensor;
-use bevy_tnua_physics_integration_layer::TnuaPipelineSystems;
-use bevy_tnua_physics_integration_layer::TnuaSystems;
 pub use spatial_ext::TnuaSpatialExtRapier2d;
 
 use self::helpers::PretendToBeRapierContext;
@@ -55,9 +53,9 @@ impl TnuaRapier2dPlugin {
 
 impl Plugin for TnuaRapier2dPlugin {
     fn build(&self, app: &mut App) {
-        app.register_required_components::<TnuaProximitySensor, Velocity>()
-            .register_required_components::<TnuaProximitySensor, ExternalForce>()
-            .register_required_components::<TnuaProximitySensor, ReadMassProperties>()
+        app.register_required_components::<TnuaMotor, Velocity>()
+            .register_required_components::<TnuaMotor, ExternalForce>()
+            .register_required_components::<TnuaMotor, ReadMassProperties>()
             .register_required_components_with::<TnuaGravity, GravityScale>(|| GravityScale(0.0));
         app.configure_sets(
             self.schedule,
@@ -78,35 +76,7 @@ impl Plugin for TnuaRapier2dPlugin {
             self.schedule,
             apply_motors_system.in_set(TnuaPipelineSystems::Motors),
         );
-        app.add_systems(
-            Update,
-            ensure_subservient_sensors_are_linked_to_rapier_context,
-        );
     }
-}
-
-fn ensure_subservient_sensors_are_linked_to_rapier_context(
-    query: Query<(Entity, &TnuaSubservientSensor), Without<RapierContextEntityLink>>,
-    links_query: Query<&RapierContextEntityLink>,
-    mut commands: Commands,
-) {
-    for (entity, subservient) in query.iter() {
-        let Ok(owner_link) = links_query.get(subservient.owner_entity) else {
-            continue;
-        };
-        commands.entity(entity).insert_if_new(*owner_link);
-    }
-}
-
-/// `bevy_rapier2d`-specific components required for Tnua to work.
-#[derive(Bundle, Default)]
-#[deprecated(
-    note = "All uses can be safely removed, components are added via bevy's required components"
-)]
-pub struct TnuaRapier2dIOBundle {
-    pub velocity: Velocity,
-    pub external_force: ExternalForce,
-    pub read_mass_properties: ReadMassProperties,
 }
 
 /// Add this component to make [`TnuaProximitySensor`] cast a shape instead of a ray.
@@ -158,31 +128,30 @@ pub(crate) fn get_collider(
 #[allow(clippy::type_complexity)]
 fn update_proximity_sensors_system(
     rapier_context_query: Query<PretendToBeRapierContext>,
-    mut query: Query<(
-        Entity,
-        &RapierContextEntityLink,
-        &GlobalTransform,
+    mut sensor_query: Query<(
         &mut TnuaProximitySensor,
+        &TnuaSensorOf,
         Option<&TnuaRapier2dSensorShape>,
         Option<&mut TnuaGhostSensor>,
-        Option<&TnuaSubservientSensor>,
+    )>,
+    owner_query: Query<(
+        &RapierContextEntityLink,
+        &GlobalTransform,
+        Option<&TnuaRapier2dSensorShape>,
         Option<&TnuaToggle>,
     )>,
     ghost_platforms_query: Query<(), With<TnuaGhostPlatform>>,
     not_platform_query: Query<(), With<TnuaNotPlatform>>,
-    other_object_query_query: Query<(&GlobalTransform, &Velocity)>,
+    other_object_query: Query<(&GlobalTransform, &Velocity)>,
 ) {
-    query.par_iter_mut().for_each(
-        |(
-            owner_entity,
-            rapier_context_entity_link,
-            transform,
-            mut sensor,
-            shape,
-            mut ghost_sensor,
-            subservient,
-            tnua_toggle,
-        )| {
+    sensor_query.par_iter_mut().for_each(
+        |(mut sensor, &TnuaSensorOf(owner_entity), shape, mut ghost_sensor)| {
+            let Ok((rapier_context_entity_link, transform, owner_shape, tnua_toggle)) =
+                owner_query.get(owner_entity)
+            else {
+                return;
+            };
+            let shape = shape.or(owner_shape);
             match tnua_toggle.copied().unwrap_or_default() {
                 TnuaToggle::Disabled => return,
                 TnuaToggle::SenseOnly => {}
@@ -204,12 +173,6 @@ fn update_proximity_sensors_system(
                 // than it should be.
                 normal: Dir3,
             }
-
-            let owner_entity = if let Some(subservient) = subservient {
-                subservient.owner_entity
-            } else {
-                owner_entity
-            };
 
             let mut query_filter = QueryFilter::new().exclude_rigid_body(owner_entity);
             let owner_solver_groups: InteractionGroups;
@@ -321,27 +284,25 @@ fn update_proximity_sensors_system(
                     // Alternative fix for https://github.com/idanarye/bevy-tnua/issues/14 - one
                     // that does not cause https://github.com/idanarye/bevy-tnua/issues/85
                     // Note that this does not solve https://github.com/idanarye/bevy-tnua/issues/87
-                    if let Some(owner_collider) = owner_collider {
-                        if owner_collider
+                    if let Some(owner_collider) = owner_collider
+                        && owner_collider
                             .shape()
                             .contains_point(&isometry, &intersection_point.into())
-                        {
-                            // I hate having to do this so much, but without it it sometimes enters
-                            // an infinte loop...
-                            cast_range_skip = proximity
-                                + if sensor.cast_range.is_finite() && 0.0 < sensor.cast_range {
-                                    0.1 * sensor.cast_range
-                                } else {
-                                    0.1
-                                };
-                            continue;
-                        };
+                    {
+                        // I hate having to do this so much, but without it it sometimes enters an
+                        // infinte loop...
+                        cast_range_skip = proximity
+                            + if sensor.cast_range.is_finite() && 0.0 < sensor.cast_range {
+                                0.1 * sensor.cast_range
+                            } else {
+                                0.1
+                            };
+                        continue;
                     }
 
                     let entity_linvel;
                     let entity_angvel;
-                    if let Ok((entity_transform, entity_velocity)) =
-                        other_object_query_query.get(entity)
+                    if let Ok((entity_transform, entity_velocity)) = other_object_query.get(entity)
                     {
                         entity_angvel = Vec3::new(0.0, 0.0, entity_velocity.angvel);
                         entity_linvel = entity_velocity.linvel.extend(0.0)